odx: support large odx entries within pdx
Entries within the pdx file were handled as byte-arrays, which
caused errors when those entries exceeded 2 GiB, due to jvm
language limitations.

switch to using InputStreams for the odx-d files to fix this.

note: this still requires large amounts of heap, due to a DOM
being required for the conversion.

+++ converter/src/main/kotlin/ODXCollection.kt
@@ -69,7 +69,7 @@ import schema.odx.UNITSPEC
 
 class ODXCollection(
     val data: Map<String, ODX>,
-    val rawSize: Int,
+    val rawSize: Long,
 ) {
     val ecuName: String by lazy {
         val ecuName =
